+++ rootdriver/state/state.py
@@ -4,16 +4,16 @@ from typing import Any
 from ..constants import DEFAULT_AGENT_STATE_AUTO_SAVE_NAME
 from ..db import JsonDB
 from ..conversation import Conversation
-from ..exception import CheckpointNotFoundError,StateDBNotFoundError
-import json
-
+from ..exceptions import CheckpointNotFoundError,StateDBNotFoundError
+from ..conversation_repo import ConversationRepo
 class State:
     def __init__(self, agent_obj: Any, db_path, auto_save: bool = False):
         self.agent = agent_obj
         self.db = JsonDB(db_path)
         self.checkpoints:dict[str,Conversation] =  {}
         self.auto_save = auto_save
-        self.auto_save_data = {"save_length": 0}
+        self.auto_save_data = {"saved_index": 0}
+        self.conversation_repo = ConversationRepo(db_path)
     def checkpoint(self, name) -> "State":
         self.checkpoints[name] = self.agent.conversation.copy()
         return self
@@ -23,8 +23,7 @@ class State:
             name: 检查点 的名称
             agent_id: agent的id， 默认当前agent的id
         """
-        agent_id = agent_id if agent_id is not None else self.agent.id
-        self.db.update(name, agent_id, self.agent.conversation.get_messages_in_list())
+        self.conversation_repo.save(self.agent.conversation, name, agent_id)
         return self.agent
 
 
@@ -32,16 +31,7 @@ class State:
         """不是全部覆写，而是计算中间差量之后追加到数据库里"""
         agent_id = agent_id if agent_id is not None else self.agent.id
 
-        current_len = len(self.agent.conversation.get_messages())
-        saved_len = self.auto_save_data["save_length"]
-
-        if current_len <= saved_len:
-            return self.agent
-        messages = self.agent.conversation.get_messages_in_list()
-
-        for i in range(saved_len, current_len):
-            self.db.append(name, agent_id, messages[i])
-        self.auto_save_data["save_length"] = current_len
+        self.conversation_repo.auto_append(self.agent.conversation, name, agent_id, self.auto_save_data["saved_index"])
         return self.agent
 
     def save_from_checkpoints(self, name: str, checkpoint_name: str, agent_id: str | None = None) -> Agent:
@@ -55,8 +45,7 @@ class State:
         checkpoint = self.checkpoints.get(checkpoint_name)
         if checkpoint is None:
             raise CheckpointNotFoundError(checkpoint_name, "memory")
-        messages = checkpoint.get_messages_in_list()
-        self.db.update(name, agent_id, messages)
+        self.conversation_repo.save(self.agent.conversation, name, agent_id)
         return self.agent
 
     def get_from_checkpoints(self, name) -> Conversation:
@@ -81,14 +70,7 @@ class State:
             Conversation
         """
         agent_id = agent_id if agent_id is not None else self.agent.id
-        try:
-            messages = self.db.read(name, agent_id, raise_error_if_not_found_agent_id=True)
-            if messages is None:
-                raise CheckpointNotFoundError(name, f"db ({self.db_path})")
-        except StateDBNotFoundError:
-            raise StateDBNotFoundError(f'agent_id = {agent_id} 数据库 {self.db.path} 中不存在')
-
-        return Conversation.from_messages_list_to_conversation(messages)
+        return self.conversation_repo.get(name, agent_id)
 
     def load_from_checkpoint(self, name:str|None=None) -> Agent:
         messages = self.get_from_checkpoints(name).messages
@@ -102,9 +84,10 @@ class State:
             agent_id: agent的id， 默认当前agent的id
         """
         agent_id = agent_id if agent_id is not None else self.agent.id
-        messages = self.get_from_db(name, agent_id).messages
-        self.agent.conversation.update_message(messages)
+        self.conversation_repo.load(self.agent.conversation, name, agent_id)
         return self.agent
+
+
 
     def update_auto_save_data_saved_length(self):
         """获取 len(self.agent.conversation.get_messages()) 更新 self.auto_save["saved_length"]的数值"""