+++ rootdriver/utils/__init__.py
@@ -1,5 +1,15 @@
 from .time import get_iso_timestamp
-from .build_message import build_user_message, build_tool_message, build_system_message, build_assistant_message, build_message
+from .build_message import (
+build_tool_message,
+build_message,
+build_user_message,
+build_assistant_message,
+build_i_message,
+build_body_message,
+build_gene_message,
+build_environment_message,
+build_system_message
+)
 from .file import ensure_file_exist
 
 __all__ = [
@@ -9,5 +19,9 @@ __all__ = [
     "build_assistant_message",
     "build_tool_message",
     "build_system_message",
+    "build_gene_message",
+    "build_environment_message",
+    "build_i_message",
+    "build_body_message",
     "ensure_file_exist"
 ]